requirements.txt
tornado
wxPython
urllib3
requests
pyinsane
